user Authorization finally completed

+++ assignments/Project_1/Time-tracker-app/delete.php
@@ -8,9 +8,10 @@ require("db.php");
 if (isset($_GET['id'])) {
 
     $id = intval($_GET['id']);
-
-    $stmt = $pdo->prepare("DELETE FROM tasks WHERE id = :id");
-    $stmt->execute([':id' => $id]);
+    $user_id = $_SESSION['user_id'];
+    
+    $stmt = $pdo->prepare("DELETE FROM tasks WHERE id = :id AND user_id = :user_id");
+    $stmt->execute([':id' => $id,':user_id' => $user_id]);
 
     header("Location: index.php");
     exit();
